+++ src/config/params.rs
@@ -1,9 +1,14 @@
+use std::cell::RefCell;
+use std::collections::HashMap;
 use std::path::Path;
+use std::rc::Rc;
+
+use once_cell::unsync::Lazy;
 
 use super::*;
 use crate::desktop::IconConfig;
 use crate::draw::{BgParams, InputTextParams, ListParams};
-use crate::font::{Font, FontBackend};
+use crate::font::{Font, FontBackend, InnerFont};
 use crate::icon::Icon;
 use crate::surface::Params as SurfaceParams;
 
@@ -12,27 +17,33 @@ macro_rules! select_conf {
         $config
             .$inner
             .$field
-            .clone()
-            .or_else(|| $config.$field.clone())
+            .as_ref()
+            .or_else(|| $config.$field.as_ref())
     };
 }
 
-impl<'a> From<&'a Config> for InputTextParams {
-    fn from(config: &'a Config) -> InputTextParams {
-        let font_color = select_conf!(config, input_text, font_color).unwrap_or(DEFAULT_FONT_COLOR);
+impl<'a> From<&'a Config> for InputTextParams<'a> {
+    fn from(config: &'a Config) -> InputTextParams<'a> {
+        let font_color = select_conf!(config, input_text, font_color)
+            .copied()
+            .unwrap_or(DEFAULT_FONT_COLOR);
 
         InputTextParams {
             font: select_conf!(config, input_text, font)
                 .map(font_by_name)
-                .unwrap_or_else(Font::default),
-            font_size: select_conf!(config, input_text, font_size).unwrap_or(DEFAULT_FONT_SIZE),
-            bg_color: select_conf!(config, input_text, bg_color).unwrap_or(DEFAULT_INPUT_BG_COLOR),
+                .unwrap_or_else(default_font),
+            font_size: select_conf!(config, input_text, font_size)
+                .copied()
+                .unwrap_or(DEFAULT_FONT_SIZE),
+            bg_color: select_conf!(config, input_text, bg_color)
+                .copied()
+                .unwrap_or(DEFAULT_INPUT_BG_COLOR),
             font_color,
             prompt_color: config.input_text.prompt_color.unwrap_or_else(|| {
                 let [r, g, b, a] = font_color.to_rgba();
                 Color::from_rgba(r, g, b, (a / 4).wrapping_mul(3))
             }),
-            prompt: config.input_text.prompt.clone(),
+            prompt: config.input_text.prompt.as_deref(),
             password: config.input_text.password,
             margin: config.input_text.margin.clone(),
             padding: config.input_text.padding.clone(),
@@ -45,9 +56,13 @@ impl<'a> From<&'a Config> for ListParams {
         ListParams {
             font: select_conf!(config, list_items, font)
                 .map(font_by_name)
-                .unwrap_or_else(Font::default),
-            font_size: select_conf!(config, list_items, font_size).unwrap_or(DEFAULT_FONT_SIZE),
-            font_color: select_conf!(config, list_items, font_color).unwrap_or(DEFAULT_FONT_COLOR),
+                .unwrap_or_else(default_font),
+            font_size: select_conf!(config, list_items, font_size)
+                .copied()
+                .unwrap_or(DEFAULT_FONT_SIZE),
+            font_color: select_conf!(config, list_items, font_color)
+                .copied()
+                .unwrap_or(DEFAULT_FONT_COLOR),
             selected_font_color: config
                 .list_items
                 .selected_font_color
@@ -97,12 +112,31 @@ impl<'a> From<&'a Config> for Option<IconConfig> {
     }
 }
 
-fn font_by_name(name: String) -> Font {
-    let path = Path::new(name.as_str());
-    if path.is_absolute() && path.exists() {
-        Font::font_by_path(path)
-    } else {
-        Font::font_by_name(name.as_str())
+fn default_font() -> Font {
+    std::thread_local! {
+        static DEFAULT_FONT: Lazy<Font> = Lazy::new(|| Rc::new(InnerFont::default()));
     }
-    .unwrap_or_else(|e| panic!("cannot find font {}: {}", name, e))
+    DEFAULT_FONT.with(|f| Rc::clone(f))
+}
+
+fn font_by_name(name: impl AsRef<str>) -> Font {
+    std::thread_local! {
+        static LOADED_FONTS: RefCell<HashMap<String, Font>> = RefCell::new(HashMap::new());
+    }
+
+    let name = name.as_ref();
+
+    if let Some(font) = LOADED_FONTS.with(|fonts| fonts.borrow().get(name).cloned()) {
+        return font;
+    }
+
+    let path = Path::new(name);
+    let font = if path.is_absolute() && path.exists() {
+        InnerFont::font_by_path(path)
+    } else {
+        InnerFont::font_by_name(name)
+    };
+    let font = Rc::new(font.unwrap_or_else(|e| panic!("cannot find font {}: {}", name, e)));
+    LOADED_FONTS.with(|fonts| fonts.borrow_mut().insert(name.to_owned(), font.clone()));
+    font
 }