+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Dropped support for font-kit backend.
 - Use native fontconfig instead of rust-fontconfig.
 
+## Fixes
+
+- Don't reload font on each ListView redraw.
+
 # 0.2.0 - 2022-10-12
 
 ## Features