+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Features
 
 - Launch multiple apps from one yofi launch based on fork (#24).
+- Nix build (#98).
 
 ## Changes